requirements.txt
streamlit
langchain
openai
python-dotenv
langchain-openai
pydantic>=2.0.0